requirements.txt
django
djangorestframework
djangorestframework-simplejwt
corsheaders
